requirements-gpu.txt
numpy==2.4.3
opencv-python==4.13.0.92
fastapi==0.135.1
python-multipart==0.0.22
uvicorn==0.41.0
onnxruntime-gpu